+++ backend/core/common/__init__.py
@@ -1,6 +1,8 @@
 from .base_models import AliasedBaseModel, ORMBase
+from .utils import get_current_datetime_utc
 
 __all__ = [
     "AliasedBaseModel",
     "ORMBase",
+    "get_current_datetime_utc",
 ]
