--- backend/core/payments/models/base.py
@@ -1,8 +1,0 @@
-from sqlalchemy.orm import DeclarativeBase
-
-from core.common import shared_metadata
-
-
-class StripeORMBase(DeclarativeBase):
-    metadata = shared_metadata
-    __table_args__ = {"schema": "stripe"}
