+++ backend/core/payments/models/stripe_event.py
@@ -9,8 +9,7 @@ from sqlalchemy import DateTime, Index, String, Text
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
-from core.common import ORMBase
-from core.common.utils import get_current_datetime_utc
+from core.common import ORMBase, get_current_datetime_utc
 
 
 @dataclass
